+++ app/src/main/AndroidManifest.xml
@@ -9,19 +9,19 @@
         android:roundIcon="@mipmap/icon_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.MainActivity">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.ScouterInitialsActivity"></activity>
-        <activity android:name=".view.SandstormActivity"></activity>
-        <activity android:name=".view.CycleActivity"></activity>
-        <activity android:name=".view.PitActivity"></activity>
-        <activity android:name=".view.EndGameActivity"></activity>
-        <activity android:name=".view.AdminActivity"></activity>
+        <activity android:name=".initials.InitialsActivity"></activity>
+        <activity android:name=".sandstorm.SandstormActivity"></activity>
+        <activity android:name=".cycle.CycleActivity"></activity>
+        <activity android:name=".pit.PitActivity"></activity>
+        <activity android:name=".endgame.EndGameActivity"></activity>
+        <activity android:name=".admin.AdminActivity"></activity>
 
         <provider
             android:name="androidx.core.content.FileProvider"
