+++ app/src/main/java/com/frc107/scouting2019/ScoutViewModel.java
@@ -1,0 +1,45 @@
+package com.frc107.scouting2019;
+
+import com.frc107.scouting2019.ScoutModel;
+
+import androidx.databinding.BaseObservable;
+
+public abstract class ScoutViewModel extends BaseObservable {
+    protected ScoutModel model;
+
+    public int getFirstUnfinishedQuestionId() {
+        return model.getFirstUnfinishedQuestionId();
+    }
+
+    public boolean isFormComplete() {
+        return model.isFormComplete();
+    }
+
+    public boolean areNoQuestionsAnswered() {
+        return model.areNoQuestionsAnswered();
+    }
+
+    public boolean setAnswer(int questionId, String answer) {
+        return model.setAnswer(questionId, answer);
+    }
+
+    public boolean setAnswer(int questionId, int answer) {
+        return model.setAnswer(questionId, answer);
+    }
+
+    public boolean setAnswer(int questionId, boolean answer) {
+        return model.setAnswer(questionId, answer);
+    }
+
+    public String getAnswerForQuestion(int questionId) {
+        return model.getAnswerForQuestion(questionId);
+    }
+
+    public String save() {
+        return model.save();
+    }
+
+    public void clearAllAnswers() {
+        model.clearAllQuestions();
+    }
+}
