+++ app/src/main/java/com/frc107/scouting2019/endgame/EndGameViewModel.java
@@ -1,0 +1,13 @@
+package com.frc107.scouting2019.endgame;
+
+import com.frc107.scouting2019.ScoutViewModel;
+
+public class EndGameViewModel extends ScoutViewModel {
+    public EndGameViewModel() {
+        model = new EndGameModel();
+    }
+
+    public String finish() {
+        return ((EndGameModel) model).finish();
+    }
+}
