--- app/src/main/java/com/frc107/scouting2019/model/InitialsModel.java
@@ -1,20 +1,0 @@
-package com.frc107.scouting2019.model;
-
-import com.frc107.scouting2019.Scouting;
-
-public class InitialsModel {
-    private String initials;
-
-    public InitialsModel() {
-        initials = "";
-    }
-
-    public void setInitials(String initals) {
-        this.initials = initals;
-        Scouting.getInstance().setInitials(initials);
-    }
-
-    public String getInitials() {
-        return initials;
-    }
-}
