--- app/src/main/java/com/frc107/scouting2019/view/MainActivity.java
@@ -1,42 +1,0 @@
-package com.frc107.scouting2019.view;
-
-import android.content.Intent;
-import android.os.Bundle;
-
-import com.frc107.scouting2019.BuildConfig;
-import com.frc107.scouting2019.R;
-import com.frc107.scouting2019.Scouting;
-
-import android.provider.Settings;
-import android.view.View;
-import android.widget.TextView;
-
-public class MainActivity extends BaseActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        int versionCode = BuildConfig.VERSION_CODE;
-        String versionName = BuildConfig.VERSION_NAME;
-
-        TextView versionTextView = findViewById(R.id.versionNum_textView);
-        versionTextView.setText(String.format("v%s - %d", versionName, versionCode));
-
-        String uniqueId = Settings.Secure.getString(getContentResolver(), Settings.Secure.ANDROID_ID);
-        Scouting.getInstance().setUniqueId(uniqueId);
-    }
-
-    public void showMatch(View view) {
-        startActivity(new Intent(this, ScouterInitialsActivity.class));
-    }
-
-    public void showPit(View view) {
-        startActivity(new Intent(this, PitActivity.class));
-    }
-
-    public void showAdmin(View view) {
-        startActivity(new Intent(this, AdminActivity.class));
-    }
-}
