--- app/src/main/java/com/frc107/scouting2019/viewmodel/InitialsViewModel.java
@@ -1,21 +1,0 @@
-package com.frc107.scouting2019.viewmodel;
-
-import com.frc107.scouting2019.model.InitialsModel;
-
-import androidx.lifecycle.ViewModel;
-
-public class InitialsViewModel extends ViewModel {
-    private InitialsModel model;
-
-    public InitialsViewModel() {
-        model = new InitialsModel();
-    }
-
-    public void setInitials(String initials) {
-        model.setInitials(initials);
-    }
-
-    public String getInitials() {
-        return model.getInitials();
-    }
-}
