--- app/src/main/java/com/frc107/scouting2019/viewmodel/SandstormViewModel.java
@@ -1,22 +1,0 @@
-package com.frc107.scouting2019.viewmodel;
-
-import com.frc107.scouting2019.model.SandstormModel;
-import com.frc107.scouting2019.model.question.Question;
-
-public class SandstormViewModel extends ScoutViewModel {
-    public SandstormViewModel() {
-        model = new SandstormModel();
-    }
-
-    public void finish() {
-        ((SandstormModel) model).finish();
-    }
-
-    public int getTeamNumber() {
-        return ((SandstormModel) model).getTeamNumber();
-    }
-
-    public boolean shouldAllowStartingPiece() {
-        return ((SandstormModel) model).shouldAllowStartingPiece();
-    }
-}
